Cambios minimos GUI
cambios minimos
